images/heic_to_jpg/requirements.txt
pillow==10.2.0
pillow_heif==0.15.0
